+++ Changes.md
@@ -4,6 +4,10 @@ This file is a manually maintained list of changes for each release. Feel free
 to add your changes here when sending pull requests. Also send corrections if
 you spot any mistakes.
 
+## HEAD
+
+* Fix `pool.query` not invoking callback on connection error #872
+
 ## v2.4.0 (2014-07-13)
 
 * Add code `POOL_NOEXIST` in PoolCluster error #846